Move WatchMode to org.gradle.internal.watch.registry
The registry should not depend on the virtual file system,
and leaving WatchMode in `vfs` causes a cyclic dependency
when doing the continuous build changes.
